Move fabric code into new fabric-common subproject
