+++ fabric-common/src/main/java/io/github/retrooper/packetevents/factory/fabric/FabricChannelInjector.java
@@ -26,7 +26,6 @@ import io.github.retrooper.packetevents.handler.PacketDecoder;
 import io.github.retrooper.packetevents.handler.PacketEncoder;
 import io.netty.channel.Channel;
 import net.fabricmc.api.EnvType;
-import net.minecraft.world.entity.player.Player;
 
 import static com.github.retrooper.packetevents.PacketEvents.DECODER_NAME;
 import static com.github.retrooper.packetevents.PacketEvents.ENCODER_NAME;
@@ -79,8 +78,8 @@ public class FabricChannelInjector implements ChannelInjector {
             return; // this channel isn't injected by packetevents
         }
         Channel ch = (Channel) channel;
-        ((PacketDecoder) ch.pipeline().get(DECODER_NAME)).player = (Player) player;
-        ((PacketEncoder) ch.pipeline().get(ENCODER_NAME)).player = (Player) player;
+        ((PacketDecoder) ch.pipeline().get(DECODER_NAME)).player = player;
+        ((PacketEncoder) ch.pipeline().get(ENCODER_NAME)).player = player;
     }
 
     @Override
