+++ fabric-common/src/main/java/io/github/retrooper/packetevents/factory/fabric/FabricLogger.java
