+++ fabric-common/src/main/resources/assets/packetevents/icon.png
