+++ fabric-common/src/main/resources/fabric.mod.json
@@ -5,7 +5,8 @@
   "name": "PacketEvents",
   "description": "",
   "authors": [
-    "retrooper"
+    "retrooper",
+    "booky10"
   ],
   "contact": {},
   "license": "GPL-3.0",
