+++ fabric/gradle.properties
@@ -2,11 +2,6 @@
 org.gradle.jvmargs=-Xmx1G
 org.gradle.parallel=true
 
-# Fabric Properties
-# check these on https://fabricmc.net/develop
-minecraft_version=26.1-pre-2
-loader_version=0.18.4
-
 publishing.skip_files=true
 publishing.platform=fabric
 publishing.modrinth.loaders=fabric
